created login page

+++ backend/controllers/budgetController.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler")
 const Budget = require("../models/budgetModel")
 
 const getBudgets = asyncHandler(async (req, res)=>{
-    const budgets = await Budget.find()
+    const budgets = await Budget.find({user_id: req.user.id})
     res.status(200).json(budgets)
 })
 
@@ -15,7 +15,8 @@ const createBudget = asyncHandler(async (req, res)=>{
     const budget = await Budget.create({
         income,
         totalBudget,
-        amountUsed
+        amountUsed,
+        user_id : req.user.id
     })
     res.status(201).json(budget)
 })
@@ -26,6 +27,11 @@ const updateBudget = asyncHandler(async (req, res)=>{
         res.status(404)
         throw new Error("Budget not found")
     }
+    if (budget.user_id.toString() != req.user.id){
+        res.status(403)
+        throw new Error("User doesn't have permission to update other users contacts")
+    }
+
     const updatedBudget = await Budget.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -39,6 +45,10 @@ const deleteBudget = asyncHandler( async (req, res)=>{
     if (!budget){
         res.status(404)
         throw new Error("Budget not found")
+    }
+    if (budget.user_id.toString() != req.user.id){
+        res.status(403)
+        throw new Error("User doesn't have permission to delete other users contacts")
     }
     await budget.deleteOne()
     res.status(200).json(budget)
